fix: Look for documents using an ObjectId.
IDs in MongoDB can be stored as either strings or Object IDs. This meant that if you created a document
and let Mongo assign an id to it, Esix wouldn't be able to find it as it was only looking for the
hex representation.

With this change, `BaseModel.find(id)` will look for documents with either a matching Object id or
a string representation of it.